src/test/java/org/example/stocktrader/manager/BarMessageQueueManagerTest.java
package org.example.stocktrader.manager;

import net.jacobpeterson.alpaca.model.endpoint.marketdata.common.realtime.bar.BarMessage;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import java.time.ZonedDateTime;
import java.util.HashMap;
import java.util.Map;
import java.util.Queue;
import static org.mockito.Mockito.*;

class BarMessageQueueManagerTest {

    private BarMessageQueueManager barMessageQueueManager;

    @Mock
    private Queue<BarMessage> mockQueue;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);

        // Mock the queue
        mockQueue = mock(Queue.class);

        // Create a mock BarMessage for peek()
        BarMessage mockPeekBarMessage = mock(BarMessage.class);
        when(mockPeekBarMessage.getTimestamp()).thenReturn(ZonedDateTime.now());
        when(mockQueue.peek()).thenReturn(mockPeekBarMessage);

        // Create a map and put the mocked queue inside it
        Map<String, Queue<BarMessage>> barMessagesInMemoryMap = new HashMap<>();
        barMessagesInMemoryMap.put("AAPL", mockQueue);

        // Initialize the BarMessageQueueManager with the mocked map
        barMessageQueueManager = new BarMessageQueueManager(barMessagesInMemoryMap);
    }

    @Test
    void testPublish() {
        BarMessage mockBarMessage = mock(BarMessage.class);
        when(mockBarMessage.getSymbol()).thenReturn("AAPL");
        when(mockBarMessage.getTimestamp()).thenReturn(ZonedDateTime.now());
        when(mockBarMessage.getClose()).thenReturn(150.0);

        barMessageQueueManager.publish(mockBarMessage);

        // Verify that the queue handling methods are called correctly
        verify(mockQueue, times(1)).add(mockBarMessage);
        verify(mockQueue, atLeastOnce()).isEmpty();
        verify(mockQueue, atLeastOnce()).peek();

        // Add more verifications and assertions as needed
    }

    // Additional test methods as required for coverage
}